refactor(consensus): move run_consensus out of lib.rs
this is in preparation for adding more functionality with the sync channel. I prefer not to
have functionality in the lib file as I see its purpose as cataloguing the modules and
public exports of the crate.
